requirements.txt
langchain==0.1.0
openai==1.3.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
streamlit==1.28.0
pypdf==3.17.1
tiktoken==0.5.1
numpy==1.24.3
pandas==2.1.1
